assign workouts to users

+++ frontend/src/components/WorkoutDetails.tsx
@@ -1,13 +1,19 @@
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }: { workout: Workout }) => {
   const { dispatch } = useWorkoutsContext();
+  const { user } = useAuthContext();
 
   async function handleDelete() {
+    if (!user) return;
     const response = await fetch(`/api/workouts/${workout._id}`, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user!.token}`,
+      },
     });
     const json = await response.json();
     if (response.ok) {
